Add function to convert HGVS position to a tuple
The idea is to have at most 3 positions for each HGVS position, and then
compare these tuples to compare the HGVS positions.

Since HGVS positions 'reset' after the coding region, e.g. 1*, this
position is converted to (math.inf, 1) to ensure it is always larger
than regular HGVS positions such as -10 or 300.